General refactoring

+++ include/matching/market.h
@@ -130,7 +130,7 @@ public:
      * @param new_price the new price to assign to the order, require that price is positive.
      * @return ErrorStatus indicating whether the order was successfully replaced.
      */
-    ErrorStatus replaceOrder(uint32_t symbol_id, uint64_t order_id, uint64_t new_order_id, uint32_t new_price);
+    ErrorStatus replaceOrder(uint32_t symbol_id, uint64_t order_id, uint64_t new_order_id, uint64_t new_price);
 
     /**
      * Executes an existing order in the market.
@@ -141,7 +141,7 @@ public:
      * @param price the price at which the order is executed, require that price is positive.
      * @return ErrorStatus indicating whether the order order was successfully executed.
      */
-    ErrorStatus executeOrder(uint32_t symbol_id, uint64_t order_id, uint64_t quantity, uint32_t price);
+    ErrorStatus executeOrder(uint32_t symbol_id, uint64_t order_id, uint64_t quantity, uint64_t price);
 
     /**
      * Executes an existing order in the market.
